+++ src/pages/home/home.js
@@ -24,6 +24,17 @@ class HomePage {
                 configs: {}
             }
         );
+
+        SettingsPlus.registerFeature(
+            GamePagesEnum.HOME,
+            'hideMaxedSkills',
+            {
+                title: 'Hide Maxed Skills',
+                subtitle: 'Hides skills that are already maxed out on the home page.',
+                enabledByDefault: true,
+                configs: {}
+            }
+        );
     }
 
     static titles = Object.freeze({
@@ -44,7 +55,7 @@ class HomePage {
         }
 
         if (!SettingsPlus.isEnabled(GamePagesEnum.HOME, 'addBuddyFarmButton')) {
-            ConsolePlus.log('Buddy Farm button is disabled in settings.');
+            ConsolePlus.debug('Buddy Farm button is disabled in settings.');
             return;
         }
 
@@ -64,6 +75,56 @@ class HomePage {
         
     };
 
+    hideMaxedSkills = (page) => {
+        if (!page?.container) {
+            new FarmRPGPlusError(
+                ErrorTypesEnum.PAGE_NOT_FOUND,
+                this.hideMaxedSkills.name,
+            );
+            return;
+        }
+
+        if (!SettingsPlus.isEnabled(GamePagesEnum.HOME, 'hideMaxedSkills')) {
+            ConsolePlus.debug('Hiding maxed skills is disabled in settings.');
+            return;
+        }
+
+        const $skillRows = $(page.container)
+            .find('.content-block-title:contains(\'My skills\')')
+            .next('.card')
+            .find('.row');
+
+        if ($skillRows.length === 0) {
+            new FarmRPGPlusError(
+                ErrorTypesEnum.ELEMENT_NOT_FOUND,
+                this.hideMaxedSkills.name,
+            );
+            return;
+        }
+
+        $skillRows.each((_index, element) => {
+            const $row = $(element);
+            
+            $row.children().each((_, skill) => {
+                const $skill = $(skill);
+
+                if ($skill.find('.progressbar-infinite').length > 0) {
+                    $skill.remove();
+                }
+            });
+
+            if ($row.children().length === 0) {
+                $row.remove();
+            } else if ($row.children().length === 1) {
+                $row.children().removeClass('col-33');
+                $row.children().addClass('col-100');
+            } else if ($row.children().length === 2) {
+                $row.children().removeClass('col-33');
+                $row.children().addClass('col-50');
+            }
+        });
+    };
+
     applyHandler = (page) => {
         if (!page?.container) {
             new FarmRPGPlusError(
@@ -75,6 +136,7 @@ class HomePage {
 
         ConsolePlus.log('Index page initialized:', page);
         this.addBuddyFarmButton(page);
+        this.hideMaxedSkills(page);
     };
 }